+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [bpmn-js-properties-panel](https://github.com/bpmn-io/bpm
 
 ___Note:__ Yet to be released changes appear here._
 
+* `DEPS`: update to `@bpmn-io/properties-panel@0.21.0`
+
 ## 1.6.1
 
 * `DEPS`: update to `@bpmn-io/properties-panel@0.20.1`